src/requirements.txt
absl-py==0.11.0
astor==0.8.0
beautifulsoup4==4.9.3
cachetools==4.1.1
cefpython3==66.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
funcy==1.15
future==0.18.2
gast==0.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
importlib-metadata==2.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyInstaller==3.2.1
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.54.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
wget==3.2
wincertstore==0.2
wrapt==1.12.1
zipp==1.2.0
